+++ src/models/Settings.php
@@ -3,7 +3,6 @@
 namespace born05\twofactorauthentication\models;
 
 use craft\base\Model;
-use craft\helpers\ConfigHelper;
 
 class Settings extends Model
 {
@@ -21,90 +20,30 @@ class Settings extends Model
     public $forceBackEnd = false;
 
     /**
-     * @var mixed The URI we should use for 2FA on the front-end.
-     *
-     * See [[ConfigHelper::localizedValue()]] for a list of supported value types.
-     * @see getVerifyPath()
+     * @var string The URI we should use for 2FA on the front-end.
      */
     public $verifyPath = '';
     
     /**
-     * @var mixed The URI we should use for 2FA settings on the front-end.
-     *
-     * See [[ConfigHelper::localizedValue()]] for a list of supported value types.
-     * @see getSettingsPath()
+     * @var string The URI we should use for 2FA settings on the front-end.
      */
     public $settingsPath = '';
 
-    // Choose between using the whitelist or blacklist! Using both will block everything!
+    // Choose between using the allow or exclude! Using both will block everything!
     /**
-     * @var mixed URIs. Exact path or regex.
-     *
-     * See [[ConfigHelper::localizedValue()]] for a list of supported value types.
-     * @see getFrontEndPathWhitelist()
+     * @var array URIs. Exact path or regex.
      */
-    public $frontEndPathWhitelist = [];
+    public $frontEndPathAllow = [];
     
     /**
-     * @var mixed URIs. Exact path or regex.
-     *
-     * See [[ConfigHelper::localizedValue()]] for a list of supported value types.
-     * @see getFrontEndPathBlacklist()
+     * @var array URIs. Exact path or regex.
      */
-    public $frontEndPathBlacklist = [];
+    public $frontEndPathExclude = [];
 
     public function rules()
     {
         return [
             [['verifyFrontEnd', 'verifyBackEnd', 'forceFrontEnd', 'forceBackEnd'], 'boolean'],
         ];
-    }
-
-    /**
-     * Returns the localized Verify Path value.
-     *
-     * @param string|null $siteHandle The site handle the value should be defined for. Defaults to the current site.
-     * @return string
-     * @see verifyPath
-     */
-    public function getVerifyPath(string $siteHandle = null): string
-    {
-        return ConfigHelper::localizedValue($this->verifyPath, $siteHandle);
-    }
-
-    /**
-     * Returns the localized Settings Path value.
-     *
-     * @param string|null $siteHandle The site handle the value should be defined for. Defaults to the current site.
-     * @return string
-     * @see settingsPath
-     */
-    public function getSettingsPath(string $siteHandle = null): string
-    {
-        return ConfigHelper::localizedValue($this->settingsPath, $siteHandle);
-    }
-
-    /**
-     * Returns the localized frontEndPathWhitelist.
-     *
-     * @param string|null $siteHandle The site handle the value should be defined for. Defaults to the current site.
-     * @return string
-     * @see frontEndPathWhitelist
-     */
-    public function getFrontEndPathWhitelist(string $siteHandle = null): array
-    {
-        return ConfigHelper::localizedValue($this->frontEndPathWhitelist, $siteHandle);
-    }
-
-    /**
-     * Returns the localized frontEndPathBlacklist.
-     *
-     * @param string|null $siteHandle The site handle the value should be defined for. Defaults to the current site.
-     * @return string
-     * @see frontEndPathBlacklist
-     */
-    public function getFrontEndPathBlacklist(string $siteHandle = null): array
-    {
-        return ConfigHelper::localizedValue($this->frontEndPathBlacklist, $siteHandle);
     }
 }